refactor(alu): remove abstraction from signature
It's generally preferable to add abstraction in function declaration so
the caller doesn't need to understand the implementation and wouldn't
need to change when the function implementation changes later.  A/C to
this approach, 'alu()' should take the whole decoded Instruction object
as an arg.

But since we are designing this project as a model/representation of the
hardware, we should imitate how the alu works in the hardware.  In the
h/w, alu takes the read data values (rd1 & rd2) and a 4-bit alu control
as input.  We are choosing to reflect this in the 'alu()' function
signature.