refactor(decoder): prefer local variable over global
The reason I chose to define the const hashmaps as global even though
they're only used inside the 'decode' function was because I thought
declaring them locally would decrease performance because it would have
to initialize the hashmap on every function call.  But recent readings
suggest the opposite is true.  Global static variables may cause cache
miss whereas local variables should not.  I am still not clear which is
preferable when, but for now I think we should go the idiomatic route:
declare the variable in the scope it is used in.